bots: Disentangle the tests from the bots
This change starts to disentangle the tests from the bots

Some changes are done to fit in with the concept of the .tasks
file, but much of the locations of images and other files have
not yet moved.

Closes #6558
