+++ bots/image-scan
@@ -1,0 +1,225 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+# This file is part of Cockpit.
+#
+# Copyright (C) 2015 Red Hat, Inc.
+#
+# Cockpit is free software; you can redistribute it and/or modify it
+# under the terms of the GNU Lesser General Public License as published by
+# the Free Software Foundation; either version 2.1 of the License, or
+# (at your option) any later version.
+#
+# Cockpit is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
+
+BASELINE_PRIORITY = 10
+
+IMAGE_REFRESH = 7
+
+DEFAULT_REFRESH = {
+    "centos-7": {
+        "triggers": [ "verify/centos-7" ]
+    },
+    "continuous-atomic": {
+        "triggers": [ "verify/continuous-atomic" ]
+    },
+    "debian-testing": {
+        "triggers": [ "verify/debian-testing" ]
+    },
+    "debian-stable": {
+        "triggers": [ "verify/debian-stable" ]
+    },
+    "fedora-25": {
+        "triggers": [
+            "verify/fedora-25",
+            "verify/fedora-atomic"
+        ]
+    },
+    "fedora-26": {
+        "triggers": [
+            "verify/fedora-26"
+        ]
+    },
+    "fedora-atomic": {
+        "triggers": [ "verify/fedora-atomic" ]
+    },
+    "fedora-testing": {
+        "triggers": [ "verify/fedora-testing" ]
+    },
+    "fedora-i386": {
+        "triggers": [ "verify/fedora-i386" ]
+    },
+    "ubuntu-1604": {
+        "triggers": [ "verify/ubuntu-1604" ]
+    },
+    "openshift": {
+        "triggers": [ "container/kubernetes",
+                      "verify/fedora-25",
+                      "verify/rhel-7" ],
+        "refresh-days": 30
+    },
+    "ipa": {
+        "triggers": [ "verify/fedora-25",
+                      "verify/rhel-7",
+                      "verify/ubuntu-1604",
+                      "verify/debian-stable" ],
+        "refresh-days": 120
+    }
+}
+
+REDHAT_STORE = "https://cockpit-11.e2e.bos.redhat.com:8493"
+
+REDHAT_REFRESH = {
+    'rhel-7': {
+        'triggers': [
+            "verify/rhel-7",
+            "verify/rhel-atomic"  # builds in rhel-7
+        ],
+        'store': REDHAT_STORE
+    },
+    'rhel-atomic': {
+        'triggers': [ "verify/rhel-atomic" ],
+        'store': REDHAT_STORE
+    }
+}
+
+# Server to tell us if we can create Red Hat images
+REDHAT_PING = "http://file.rdu.redhat.com"
+
+import argparse
+import os
+import json
+import pipes
+import sys
+import time
+import urllib
+
+sys.dont_write_bytecode = True
+
+import github
+
+def main():
+    parser = argparse.ArgumentParser(description='Bot: scan for images that need updating')
+    parser.add_argument('-d', '--dry', action="store_true", default=False,
+                        help='Don''t actually change anything on GitHub')
+    parser.add_argument('-v', '--verbose', action="store_true", default=False,
+                        help="Print verbose information")
+    opts = parser.parse_args()
+    api = github.GitHub()
+
+    try:
+        results = scan(api, not opts.dry, opts.verbose)
+    except RuntimeError, ex:
+        sys.stderr.write("image-scan: " + str(ex) + "\n")
+        return 1
+
+    for result in results:
+        if result:
+            sys.stdout.write(result + "\n")
+
+    return 0
+
+# Prepare an image refresh command
+def image_refresh(priority, name, image, issue=None, triggers=[], store=None):
+    if priority <= 0:
+        return
+    cmd = "PRIORITY={priority:04d} TEST_NAME='{name}' bots/image-refresh"
+    if issue:
+        cmd += " --issue='{issue}'"
+    if store:
+        cmd += " --store='{store}'"
+    for trigger in triggers:
+        cmd += " --triggers='{0}'".format(pipes.quote(trigger))
+    cmd += " '{image}'"
+    return cmd.format(**{
+        "priority": int(priority),
+        "name": pipes.quote(name),
+        "image": pipes.quote(image),
+        "issue": pipes.quote(str(issue)),
+        "store": pipes.quote(str(store)),
+    })
+
+def scan_image_wait_times(api, policy):
+    issues = api.get("issues?labels=bot&filter=all&state=all")
+
+    wait_times = { }
+    for image, config in policy.items():
+        wait_times[image] = 0
+        for issue in issues:
+            if issue['title'] == github.ISSUE_TITLE_IMAGE_REFRESH.format(image):
+                age = time.time() - time.mktime(time.strptime(issue['created_at'], "%Y-%m-%dT%H:%M:%SZ"))
+                refresh_days = config.get("refresh-days", IMAGE_REFRESH)
+                wait_time = refresh_days - (age / (24 * 60 * 60))
+                if wait_time > wait_times[image]:
+                    wait_times[image] = wait_time
+    return wait_times
+
+def scan_for_image_tasks(api, policy):
+    results = [ ]
+    requested = set()
+
+    # Trigger on explicit requests
+    def issue_requests_image_refresh(issue, comments, image):
+        request = "bot: " + github.ISSUE_TITLE_IMAGE_REFRESH.format(image)
+        in_process_prefix = "Image creation for {} in process".format(image)
+
+        needed = False
+        for body in [ issue['body'] ] + [ c['body'] for c in comments ]:
+            if body == request:
+                needed = True
+            if body.startswith(in_process_prefix):
+                needed = False
+        return needed
+
+    for issue in api.get("issues?labels=bot&state=open"):
+        comments = api.get(issue['comments_url'])
+        for image in policy:
+            if issue_requests_image_refresh(issue, comments, image):
+                requested.add(image)
+                store = policy.get(image, { }).get("store", None)
+                triggers = policy.get(image, { }).get("triggers", [ ])
+                results.append(image_refresh(BASELINE_PRIORITY, "refresh-" + image, image, issue, triggers, store))
+
+    # Trigger based on how old the youngest issue is
+    for image, wait_time in scan_image_wait_times(api, policy).items():
+        if wait_time <= 0 and image not in requested:
+            store = policy.get(image, { }).get("store", None)
+            triggers = policy.get(image, { }).get("triggers", [ ])
+            results.append(image_refresh(BASELINE_PRIORITY, "refresh-" + image, image, None, triggers, store))
+
+    return results
+
+def scan(api, update, verbose):
+    kvm = os.access("/dev/kvm", os.R_OK | os.W_OK)
+    if not kvm:
+        sys.stderr.write("tests-scan: No /dev/kvm access, not creating images here\n")
+        return []
+
+    policy = DEFAULT_REFRESH
+
+    # Check if we have access to Red Hat network
+    try:
+        urllib.urlopen(REDHAT_PING).read()
+        policy.update(REDHAT_REFRESH)
+    except IOError:
+        pass
+
+    task_entries = scan_for_image_tasks(api, policy)
+
+    if verbose:
+        for image, wait_time in scan_image_wait_times(api, policy).items():
+            # Images with wait_time <= 0 are already included in the list
+            # of tasks above, so we don't need to list them here.
+            if wait_time > 0:
+                sys.stdout.write("# {0} will be refreshed in {1:.1f} days.\n".format(image, wait_time))
+
+    return task_entries
+
+if __name__ == '__main__':
+    sys.exit(main())
