+++ test/README
@@ -96,7 +96,7 @@ You can set these environment variables to configure the test suite:
                 "fedora-testing"
                 "rhel-7"
                 "ubuntu-1604"
-             "fedora-25" is the default (testinfra.py)
+             "fedora-25" is the default (testvm.py)
 
   TEST_DATA  Where to find and store test machine images.  The
              default is the same directory that this README file is in.
