--- test/avocado/testimagetask.py
@@ -1,1 +1,0 @@
-../common/testimagetask.py
