--- test/avocado/testinfra.py
@@ -1,1 +1,0 @@
-../common/testinfra.py
