--- test/avocado/testpulltask.py
@@ -1,1 +1,0 @@
-../common/testpulltask.py
