+++ test/verify/check-embed
@@ -20,7 +20,6 @@
 
 import parent
 from testlib import *
-import testinfra
 
 class TestEmbed(MachineCase):
     def testBasic(self):
@@ -29,7 +28,7 @@ class TestEmbed(MachineCase):
 
         m.start_cockpit()
 
-        b.open("file://%s/verify/files/embed-cockpit.html" % testinfra.TEST_DIR)
+        b.open("file://%s/verify/files/embed-cockpit.html" % TEST_DIR)
         b.wait_present("#embed-loaded")
         b.set_val("#embed-address", "http://" + m.address)
         b.click("#embed-full")
