+++ test/verify/testsuite-prepare
@@ -17,6 +17,7 @@
 # along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
 
 import argparse
+import errno
 import os
 import sys
 import glob
@@ -24,7 +25,8 @@ import shutil
 import subprocess
 
 import parent
-from common import testinfra
+from common import testlib
+from common import testvm
 from common import vmmanage
 
 if __name__ == "__main__":
@@ -35,7 +37,7 @@ if __name__ == "__main__":
     parser.add_argument('-v', '--verbose', action='store_true', help='Display verbose progress details')
     parser.add_argument('-n', '--noinstall', action='store_true', help='Do not install cockpit to base image')
     parser.add_argument('-f', '--force', action='store_true', help='Force update of images')
-    parser.add_argument('image', nargs='?', default=testinfra.DEFAULT_IMAGE, help='The image to use')
+    parser.add_argument('image', nargs='?', default=testvm.DEFAULT_IMAGE, help='The image to use')
     args = parser.parse_args()
 
     test_os = args.image
@@ -43,7 +45,11 @@ if __name__ == "__main__":
 
     os.system("vm-reset")
     try:
-        subprocess.check_call(["vm-download", build_os])
+        subprocess.check_call(["vm-download", build_os, test_os])
+    except OSError, ex:
+        if ex.errno != errno.ENOENT:
+            raise
+        sys.stderr.write("testsuite-prepare: missing tools to download images\n")
     except subprocess.CalledProcessError as e:
         print "unable to download all necessary images", e
         sys.exit(1)
@@ -63,6 +69,6 @@ if __name__ == "__main__":
                 os.makedirs(os.path.join(attachement_dir, "build-results"))
             except Exception as e:
                 pass
-            for cpfile in glob.glob(os.path.join(testinfra.TEST_DIR,"tmp","build-results","*.log")):
+            for cpfile in glob.glob(os.path.join(testvm.TEST_DIR,"tmp","build-results","*.log")):
                 shutil.copy(cpfile, os.path.join(attachement_dir ,"build-results"))
     sys.exit(retvalue)
