+++ test/vm-customize
@@ -22,7 +22,6 @@ import re
 import sys
 import subprocess
 
-from common import testinfra
 from common import testvm
 
 parser = argparse.ArgumentParser(
@@ -39,7 +38,7 @@ parser.add_argument('-s', '--script', action='append', dest="scriptlist",
                                       default=[], help='Run selected script inside virtual machine')
 parser.add_argument('-u', '--upload', action='append', dest="uploadlist",
                     default=[], help='Upload file/dir to destination file/dir separated by ":" example: -u file.txt:/var/lib')
-parser.add_argument('image', nargs='?', default=testinfra.DEFAULT_IMAGE, help='The image to use')
+parser.add_argument('image', nargs='?', default=testvm.DEFAULT_IMAGE, help='The image to use')
 args = parser.parse_args()
 
 def run_command(machine_instance, commandlist):
@@ -64,7 +63,6 @@ def upload_files(machine_instance, uploadfiles):
     """Upload files/directories inside image"""
     for foo in uploadfiles:
         srcfile, dest = foo.split(":")
-        # machine_instance.upload tries to interpret relative paths in relation to testinfra.TEST_DIR
         src_absolute = os.path.join(os.getcwd(), srcfile)
         machine_instance.upload([src_absolute], dest)
